# docs/app/swagger.yaml
basePath: /api/v1
definitions:
...
  management.User:
    properties:
      app_metadata:
        additionalProperties: true
        description: |-
          AppMetadata holds data that the user has read-only access to.
          For example roles, permissions, vip, etc.
          NOTE: Roles added to AppMetadata are not integrated with Auth0 Role-Based Access Control (RBAC).
          For RBAC, see the functions User.Roles, User.AssignRoles, and User.RemoveRoles.
        type: object
      blocked:
        description: True if the user is blocked from the application, false if the
          user is enabled.
        type: boolean
      client_id:
        description: Auth0 client ID. Only valid when updating email address.
        type: string
      connection:
        description: The connection the user belongs to.
        type: string
      created_at:
        description: The time the user was created.
        type: string
      description:
        description: The user-defined UTF-8 string describing their account.
        type: string
      email:
        description: The users' email.
        type: string
      family_name:
        description: The users' family name.
        type: string
      given_name:
        description: The users' given name.
        type: string
      identities:
        description: Identities is a list of user identities for when accounts are
          linked.
        items:
          $ref: '#/definitions/management.UserIdentity'
        type: array
      last_ip:
        description: Last IP address from which this user logged in. Read only, cannot
          be modified.
        type: string
      last_login:
        description: The last time the user has logged in.
        type: string
      last_password_reset:
        description: |-
          The last time the user had their password reset.
          Only available for Database connection users.
        type: string
      location:
        description: The user-defined location for this account’s profile.
        type: string
      logins_count:
        description: Total number of logins this user has performed. Read only, cannot
          be modified.
        type: integer
      multifactor:
        description: List of multi-factor authentication providers with which this
          user has enrolled.
        items:
          type: string
        type: array
      name:
        description: The users' name.
        type: string
      nickname:
        description: The users' nickname.
        type: string
      password:
        description: The users' password (mandatory for non SMS connections)
        type: string
      phone_number:
        description: |-
          The users' phone number (following the E.164 recommendation).
          Only valid for users to be added to SMS connections.
        type: string
      phone_verified:
        description: |-
          True if the user's phone number is verified, false otherwise. When the
          user is added to an SMS connection, they will not receive a verification
          SMS if this is true.
        type: boolean
      picture:
        description: The user's picture url.
        type: string
      screen_name:
        description: The screen name, handle, or alias that this user identifies themselves
          with.
        type: string
      updated_at:
        description: The last time the user was updated.
        type: string
      url:
        description: A URL provided by the user in association with their profile.
        type: string
      user_id:
        description: The users' identifier.
        type: string
      user_metadata:
        additionalProperties: true
        description: |-
          UserMetadata holds data that the user has read/write access to.
          For example color_preference, blog_url, etc.
        type: object
      username:
        description: The users' username. Only valid if the connection requires a
          username.
        type: string
      verify_email:
        description: |-
          If true, the user will receive a verification email after creation, even
          if created with email_verified set to true. If false, the user will not
          receive a verification email, even if created with email_verified set to
          false. If unspecified, defaults to the behavior determined by the value
          of email_verified.
        type: boolean
    type: object
...
  placio-app_Dto.ErrorDTO:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  placio-app_Dto.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  placio-app_ent.Like:
    properties:
      createdAt:
        description: CreatedAt holds the value of the "createdAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.LikeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the LikeQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      like:
        description: True for like, False for dislike
        type: boolean
      updatedAt:
        description: UpdatedAt holds the value of the "updatedAt" field.
        type: string
    type: object
  placio-app_ent.LikeEdges:
    properties:
      media:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Media'
        description: The media content that was liked/disliked.
      post:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Post'
        description: Post holds the value of the post edge.
      review:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Review'
        description: The review that was liked/disliked.
      user:
        allOf:
        - $ref: '#/definitions/placio-app_ent.User'
        description: The user who liked/disliked the review/media.
    type: object
  placio-app_ent.Media:
    properties:
      CreatedAt:
        description: CreatedAt holds the value of the "CreatedAt" field.
        type: string
      MediaType:
        description: image, gif, or video
        type: string
      URL:
        description: URL holds the value of the "URL" field.
        type: string
      UpdatedAt:
        description: UpdatedAt holds the value of the "UpdatedAt" field.
        type: string
      dislikeCount:
        description: Count of dislikes for this media.
        type: integer
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.MediaEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MediaQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      likeCount:
        description: Count of likes for this media.
        type: integer
    type: object
  placio-app_ent.MediaEdges:
    properties:
      categories:
        description: Categories holds the value of the categories edge.
        items:
          $ref: '#/definitions/placio-app_ent.Category'
        type: array
      post:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Post'
        description: Post holds the value of the post edge.
      review:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Review'
        description: Review holds the value of the review edge.
    type: object
...
  placio-app_ent.Place:
    properties:
      additional_info:
        additionalProperties: true
        description: AdditionalInfo holds the value of the "additional_info" field.
        type: object
      availability:
        additionalProperties: true
        description: Availability holds the value of the "availability" field.
        type: object
      city:
        description: City holds the value of the "city" field.
        type: string
      country:
        description: Country holds the value of the "country" field.
        type: string
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      description:
        description: Description holds the value of the "description" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.PlaceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PlaceQuery when eager-loading is set.
      email:
        description: Email holds the value of the "email" field.
        type: string
      features:
        description: Features holds the value of the "features" field.
        items:
          type: string
        type: array
      id:
        description: ID of the ent.
        type: string
      images:
        description: Images holds the value of the "images" field.
        items:
          type: string
        type: array
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      opening_hours:
        additionalProperties: true
        description: OpeningHours holds the value of the "opening_hours" field.
        type: object
      payment_options:
        additionalProperties: true
        description: PaymentOptions holds the value of the "payment_options" field.
        type: object
      phone:
        description: Phone holds the value of the "phone" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      place_settings:
        additionalProperties: true
        description: PlaceSettings holds the value of the "place_settings" field.
        type: object
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      social_media:
        additionalProperties: true
        description: SocialMedia holds the value of the "social_media" field.
        type: object
      special_offers:
        description: SpecialOffers holds the value of the "special_offers" field.
        type: string
      state:
        description: State holds the value of the "state" field.
        type: string
      sustainability_score:
        description: SustainabilityScore holds the value of the "sustainability_score"
          field.
        type: number
      tags:
        description: Tags holds the value of the "tags" field.
        items:
          type: string
        type: array
      type:
        description: Type holds the value of the "type" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
  placio-app_ent.PlaceEdges:
    properties:
      amenities:
        description: Amenities holds the value of the amenities edge.
        items:
          $ref: '#/definitions/placio-app_ent.Amenity'
        type: array
      bookings:
        description: Bookings holds the value of the bookings edge.
        items:
          $ref: '#/definitions/placio-app_ent.Booking'
        type: array
      business:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Business'
        description: Business holds the value of the business edge.
      categories:
        description: Categories holds the value of the categories edge.
        items:
          $ref: '#/definitions/placio-app_ent.Category'
        type: array
      categoryAssignments:
        description: CategoryAssignments holds the value of the categoryAssignments
          edge.
        items:
          $ref: '#/definitions/placio-app_ent.CategoryAssignment'
        type: array
      events:
        description: Events holds the value of the events edge.
        items:
          $ref: '#/definitions/placio-app_ent.Event'
        type: array
      faqs:
        description: Faqs holds the value of the faqs edge.
        items:
          $ref: '#/definitions/placio-app_ent.FAQ'
        type: array
      followerUsers:
        description: FollowerUsers holds the value of the followerUsers edge.
        items:
          $ref: '#/definitions/placio-app_ent.UserFollowPlace'
        type: array
      likedByUsers:
        description: LikedByUsers holds the value of the likedByUsers edge.
        items:
          $ref: '#/definitions/placio-app_ent.UserLikePlace'
        type: array
      menus:
        description: Menus holds the value of the menus edge.
        items:
          $ref: '#/definitions/placio-app_ent.Menu'
        type: array
      ratings:
        description: Ratings holds the value of the ratings edge.
        items:
          $ref: '#/definitions/placio-app_ent.Rating'
        type: array
      reservations:
        description: Reservations holds the value of the reservations edge.
        items:
          $ref: '#/definitions/placio-app_ent.Reservation'
        type: array
      reviews:
        description: Reviews holds the value of the reviews edge.
        items:
          $ref: '#/definitions/placio-app_ent.Review'
        type: array
      rooms:
        description: Rooms holds the value of the rooms edge.
        items:
          $ref: '#/definitions/placio-app_ent.Room'
        type: array
      users:
        description: Users holds the value of the users edge.
        items:
          $ref: '#/definitions/placio-app_ent.User'
        type: array
    type: object
  placio-app_ent.Post:
    properties:
      Content:
        description: Content holds the value of the "Content" field.
        type: string
      CreatedAt:
        description: CreatedAt holds the value of the "CreatedAt" field.
        type: string
      Privacy:
        allOf:
        - $ref: '#/definitions/placio-app_ent_post.Privacy'
        description: Privacy holds the value of the "Privacy" field.
      UpdatedAt:
        description: UpdatedAt holds the value of the "UpdatedAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.PostEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PostQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
    type: object
...
  placio-app_ent.Review:
    properties:
      content:
        description: User's review to the business/place/event.
        type: string
      createdAt:
        description: When the review was created.
        type: string
      dislikeCount:
        description: Count of dislikes for this review.
        type: integer
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.ReviewEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ReviewQuery when eager-loading is set.
      flag:
        description: Flag for this review.
        type: string
      id:
        description: ID of the ent.
        type: string
      likeCount:
        description: Count of likes for this review.
        type: integer
      score:
        description: Score should be between 1 and 5.
        type: number
    type: object
...
  placio-app_ent.User:
    properties:
      app_settings:
        additionalProperties: true
        description: AppSettings holds the value of the "app_settings" field.
        type: object
      auth0_data:
        allOf:
        - $ref: '#/definitions/management.User'
        description: Auth0Data holds the value of the "auth0_data" field.
      auth0_id:
        description: Auth0ID holds the value of the "auth0_id" field.
        type: string
      bio:
        description: Bio holds the value of the "bio" field.
        type: string
      cover_image:
        description: CoverImage holds the value of the "cover_image" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.UserEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
      latitude:
        description: Latitude holds the value of the "latitude" field.
        type: string
      location:
        description: Location holds the value of the "location" field.
        type: string
      longitude:
        description: Longitude holds the value of the "longitude" field.
        type: string
      map_coordinates:
        additionalProperties: true
        description: MapCoordinates holds the value of the "map_coordinates" field.
        type: object
      name:
        description: Name holds the value of the "name" field.
        type: string
      picture:
        description: Picture holds the value of the "picture" field.
        type: string
      relevance_score:
        description: RelevanceScore holds the value of the "relevance_score" field.
        type: number
      search_text:
        description: SearchText holds the value of the "search_text" field.
        type: string
      user_settings:
        additionalProperties: true
        description: UserSettings holds the value of the "user_settings" field.
        type: object
      username:
        description: Username holds the value of the "username" field.
        type: string
      website:
        description: Website holds the value of the "website" field.
        type: string
    type: object
...
  placio-app_ent.UserLikePlace:
    properties:
      CreatedAt:
        description: CreatedAt holds the value of the "CreatedAt" field.
        type: string
      UpdatedAt:
        description: UpdatedAt holds the value of the "UpdatedAt" field.
        type: string
      edges:
        allOf:
        - $ref: '#/definitions/placio-app_ent.UserLikePlaceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the UserLikePlaceQuery when eager-loading is set.
      id:
        description: ID of the ent.
        type: string
    type: object
  placio-app_ent.UserLikePlaceEdges:
    properties:
      place:
        allOf:
        - $ref: '#/definitions/placio-app_ent.Place'
        description: Place holds the value of the place edge.
      user:
        allOf:
        - $ref: '#/definitions/placio-app_ent.User'
        description: User holds the value of the user edge.
    type: object
...
host: https://api.palnight.com
info:
  contact:
    email: support@placio.io
    name: Placio Ltd
    url: https://placio.io
  description: This is the documentation for the Placio Application Api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://placio.io/terms
  title: Placio Application Api
  version: "0.01"
paths:
...
  /api/v1/likes/{likeID}:
    delete:
      consumes:
      - application/json
      description: Remove a like from a post
      parameters:
      - description: ID of the like
        in: path
        name: likeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unliked post
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Unlike a post
      tags:
      - Like
  /api/v1/likes/place/{placeID}:
    get:
      consumes:
      - application/json
      description: Retrieve all likes for a place
      parameters:
      - description: ID of the place
        in: path
        name: placeID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of UserLikePlace records
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.UserLikePlace'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get place likes
      tags:
      - Like
  /api/v1/likes/place/{userID}/place/{placeID}:
    post:
      consumes:
      - application/json
      description: Allows a user to like a specific place
      parameters:
      - description: ID of the user
        in: path
        name: userID
        required: true
        type: string
      - description: ID of the place
        in: path
        name: placeID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Successfully liked place
          schema:
            $ref: '#/definitions/placio-app_ent.UserLikePlace'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Like a place
      tags:
      - Like
  /api/v1/likes/place/{userLikePlaceID}:
    delete:
      consumes:
      - application/json
      description: Allows a user to unlike a specific place
      parameters:
      - description: ID of the UserLikePlace record
        in: path
        name: userLikePlaceID
        required: true
        type: string
      - description: Provide JWT access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unliked place
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Unlike a place
      tags:
      - Like
  /api/v1/likes/post/{postID}:
    get:
      consumes:
      - application/json
      description: Retrieve all likes for a post
      parameters:
      - description: ID of the post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of likes
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Like'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get post likes
      tags:
      - Like
  /api/v1/likes/user/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieve all likes by a user
      parameters:
      - description: ID of the user
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of likes
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.Like'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get user likes
      tags:
      - Like
  /api/v1/likes/user/{userID}/post/{postID}:
    post:
      consumes:
      - application/json
      description: Add a like to a post by a user
      parameters:
      - description: ID of the user
        in: path
        name: userID
        required: true
        type: string
      - description: ID of the post
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully liked post
          schema:
            $ref: '#/definitions/placio-app_Dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Like a post
      tags:
      - Like
  /api/v1/likes/user/place/{userID}:
    get:
      consumes:
      - application/json
      description: Retrieve all places liked by a user
      parameters:
      - description: ID of the user
        in: path
        name: userID
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of UserLikePlace records
          schema:
            items:
              $ref: '#/definitions/placio-app_ent.UserLikePlace'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/placio-app_Dto.ErrorDTO'
      summary: Get user liked places
      tags:
      - Like
...
schemes:
- http
- https
swagger: "2.0"